# fitness-cms/web-cms/public/swagger.yaml
openapi: 3.0.3
info:
  title: FitToday CMS API
  description: |
    API for the FitToday fitness CMS platform. Connects personal trainers with students
    through programs, workouts, and real-time communication.

    ## Authentication
    All protected endpoints require a Firebase Auth token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```

    ## Roles
    - **trainer** - Personal trainers who create programs and manage students
    - **student** - Students who follow programs and track progress
    - **admin** - Platform administrators

    ## Plans
    | Plan    | Price (BRL) | Commission | Max Programs | Max Students |
    |---------|-------------|------------|--------------|--------------|
    | Starter | Free        | 10%        | 3            | 5            |
    | Pro     | R$97/mo     | 5%         | Unlimited    | Unlimited    |
    | Elite   | R$197/mo    | 0%         | Unlimited    | Unlimited    |
  version: 1.0.0
  contact:
    name: FitToday Engineering
servers:
  - url: https://fittoday.vercel.app
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID token
  schemas:
...
    Program:
      type: object
      properties:
        id:
          type: string
        trainerId:
          type: string
        studentId:
          type: string
        title:
          type: string
        description:
          type: string
        coverImageURL:
          type: string
        previewVideoURL:
          type: string
        workoutPdfUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        duration:
          type: object
          properties:
            weeks:
              type: integer
            daysPerWeek:
              type: integer
            avgSessionMinutes:
              type: integer
        requirements:
          type: object
          properties:
            equipment:
              type: array
              items:
                type: string
            fitnessLevel:
              type: string
            prerequisites:
              type: string
        content:
          type: object
          properties:
            totalWorkouts:
              type: integer
            totalExercises:
              type: integer
            includesNutrition:
              type: boolean
            includesSupplements:
              type: boolean
            hasVideoGuides:
              type: boolean
        stats:
          type: object
          properties:
            totalSales:
              type: integer
            activeStudents:
              type: integer
            completionRate:
              type: number
            averageRating:
              type: number
            totalReviews:
              type: integer
        weeks:
          type: array
          items:
            type: object
        status:
          type: string
          enum: [draft, published, archived]
        visibility:
          type: string
          enum: [public, private]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
    CreateProgramRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string
        studentId:
          type: string
        coverImageURL:
          type: string
        previewVideoURL:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          default: intermediate
        duration:
          type: object
          properties:
            weeks:
              type: integer
              default: 8
            daysPerWeek:
              type: integer
              default: 4
            avgSessionMinutes:
              type: integer
              default: 60
        requirements:
          type: object
          properties:
            equipment:
              type: array
              items:
                type: string
            prerequisites:
              type: string
        content:
          type: object
          properties:
            totalWorkouts:
              type: integer
            totalExercises:
              type: integer
            includesNutrition:
              type: boolean
            includesSupplements:
              type: boolean
            hasVideoGuides:
              type: boolean
        weeks:
          type: array
          items:
            type: object
        status:
          type: string
          enum: [draft, published]
          default: draft
        visibility:
          type: string
          enum: [public, private]
          default: public

    # ---- Exercises ----
...
paths:
...
  /api/programs/{id}:
    get:
      tags: [Programs]
      summary: Get a program
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Program not found

    put:
      tags: [Programs]
      summary: Update a program
      description: Trainer-only. Updates only provided fields. Supports multipart for file uploads.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string with fields to update
                cover:
                  type: string
                  format: binary
                pdf:
                  type: string
                  format: binary
                video:
                  type: string
                  format: binary
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "200":
          description: Program updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  coverImageURL:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
        "401":
          description: Unauthorized
        "403":
          description: Not the program owner
        "404":
          description: Program not found

    delete:
      tags: [Programs]
      summary: Archive a program (soft delete)
      description: |
        Trainer-only. Sets status to "archived" and visibility to "private".
        Cannot delete programs with active students.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [archived]
        "400":
          description: Cannot delete with active students
        "401":
          description: Unauthorized
        "403":
          description: Not the program owner
        "404":
          description: Program not found
...
  /api/stripe/webhook:
    post:
      tags: [Stripe]
      summary: Stripe webhook handler
      security: []
      description: |
        Handles Stripe webhook events. Requires stripe-signature header for verification.

        **Handled events:**
        - `account.updated` - Update trainer's Stripe account status
        - `checkout.session.completed` - Process program purchase or platform subscription
        - `invoice.paid` / `invoice.payment_succeeded` - Renew subscriptions
        - `invoice.payment_failed` - Mark subscriptions as past_due
        - `customer.subscription.updated` - Sync subscription status
        - `customer.subscription.deleted` - Cancel subscriptions and revoke access
        - `transfer.created` - Update trainer balance
        - `charge.refunded` - Process refunds
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        "400":
          description: Invalid signature

  # ================================================================
  # ADMIN
  # ================================================================
...